Move text completions API endpoints to own module
